refactor: DOMO及びフォロー関連処理をユーティリティモジュールに分割
- DOMO操作を行う `domo_activity` 関数を `domo_utils.py` に分割。
- リストアイテムからのフォロー操作（ボタン検索、クリック検証）を行う関数群を `follow_utils.py` に分割。
- `yamap_auto_domo.py` からこれらの関数を呼び出すように修正。
- `TASK_MANAGEMENT.md` を更新。
